use correct lucene geo queries for different index versions
use temoraryFolder rule for temp directories

split LuceneQueryBuilderTest

added within tests with various index versions

added changelog entry
